truncate metal before comparing

+++ src/deep_blue/platform/logic/command.py
@@ -173,7 +173,7 @@ class Produce(Command):
             return False
         for element in game.map_info.elements.values():
             if isinstance(element, Base) and element.team == team:
-                if element.metal < PROPERTY[self.kind]['health_max'] * METAL_PER_HEALTH:
+                if element.metal < int(PROPERTY[self.kind]['health_max'] * METAL_PER_HEALTH):
                     return False
         game.commands[team].append(self)
         self.team = team
